requirements.txt
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cmake==3.27.2
filelock==3.12.3
flwr==1.4.0
frozenlist==1.4.0
grpcio==1.57.0
hydra-core==1.3.2
idna==3.4
iterators==0.0.2
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
lit==16.0.6
MarkupSafe==2.1.3
mpmath==1.3.0
msgpack==1.0.5
networkx==3.1
numpy==1.25.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
omegaconf==2.3.0
packaging==23.1
Pillow==10.0.0
protobuf==3.20.3
PyYAML==6.0.1
ray==2.6.3
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.0
sympy==1.12
torch==2.0.1
torchvision==0.15.2
triton==2.0.0
typing_extensions==4.7.1
urllib3==2.0.4
